Final changes for 0.11.4

+++ rdiff-backup/testing/metadatatest.py
@@ -8,7 +8,7 @@ class MetadataTest(unittest.TestCase):
 	def make_temp(self):
 		"""Make temp directory testfiles/output"""
 		global tempdir
-		tempdir.delete()
+		if tempdir.lstat(): tempdir.delete()
 		tempdir.mkdir()
 
 	def testQuote(self):
